fix: 账号备份详情页显示sftp服务器信息
